refac: add function to generate profile embeds
